services/rayonPostling/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-login==0.6.2
pillow==10.0.0
cryptography==41.0.3
redis==5.0.1
gunicorn==21.2.0
werkzeug==2.3.7
sqlalchemy==2.0.19
pycryptodome==3.19.0
